requirements.txt
TTS
termcolor
nats-py
boto3
python-dotenv